Deltatime en segundos
- Se ahorran decenas de divisiones entre 1000
para cambiar de milisegundos->segundos

+++ src/Kepri3D/Component.h
@@ -26,7 +26,7 @@ public:
 	inline void setEntity(Entity* e) { this->entity = e; }
 
 	/* Método virtual puro que debe implementar cada tipo de componente */
-	virtual void update(GLuint deltaTime) = 0;
+	virtual void update(float deltaTime) = 0;
 
 	/* Callback para cuando la entidad colisione con otra */
 	virtual void onCollision(Collider* other) {}
